+++ CHANGELOG.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+.. _v44-0-1:
+
+44.0.1 - 2025-02-11
+~~~~~~~~~~~~~~~~~~~
+
+* We now build ``armv7l`` ``manylinux`` wheels and publish them to PyPI.
+
 .. _v44-0-0:
 
 44.0.0 - 2024-11-27
@@ -25,7 +32,7 @@ Changelog
   when using OpenSSL 3.2.0+.
 * Added support for the :class:`~cryptography.x509.Admissions` certificate extension.
 * Added basic support for PKCS7 decryption (including S/MIME 3.2) via
-  :func:`~cryptography.hazmat.primitives.serialization.pkcs7.pkcs7_decrypt_der`, 
+  :func:`~cryptography.hazmat.primitives.serialization.pkcs7.pkcs7_decrypt_der`,
   :func:`~cryptography.hazmat.primitives.serialization.pkcs7.pkcs7_decrypt_pem`, and
   :func:`~cryptography.hazmat.primitives.serialization.pkcs7.pkcs7_decrypt_smime`.